+++ scripts/test/Selenium/Agent/AgentTicketZoom.t
@@ -21,6 +21,25 @@ $Selenium->RunTest(
         # get helper object
         my $Helper = $Kernel::OM->Get('Kernel::System::UnitTest::Helper');
 
+        # Overload CustomerUser => Map setting defined in the Defaults.pm.
+        my $DefaultCustomerUser = $Kernel::OM->Get('Kernel::Config')->Get("CustomerUser");
+        $DefaultCustomerUser->{Map}->[5] = [
+            'UserEmail',
+            'Email',
+            'email',
+            1,
+            1,
+            'var',
+            '[% Env("CGIHandle") %]?Action=AgentTicketCompose;ResponseID=1;TicketID=[% Data.TicketID | uri %];ArticleID=[% Data.ArticleID | uri %]',
+            0,
+            '',
+            'AsPopup OTRSPopup_TicketAction',
+        ];
+        $Helper->ConfigSettingChange(
+            Key   => 'CustomerUser',
+            Value => $DefaultCustomerUser,
+        );
+
         # make sure we start with RuntimeDB search
         $Helper->ConfigSettingChange(
             Valid => 1,
@@ -60,19 +79,37 @@ $Selenium->RunTest(
         my $TitleRandom  = "Title" . $Helper->GetRandomID();
         my $TicketNumber = $TicketObject->TicketCreateNumber();
         my $TicketID     = $TicketObject->TicketCreate(
-            TN         => $TicketNumber,
-            Title      => $TitleRandom,
-            Queue      => 'Raw',
-            Lock       => 'unlock',
-            Priority   => '3 normal',
-            State      => 'open',
-            CustomerID => $TestCustomerUserID{UserCustomerID},
-            OwnerID    => 1,
-            UserID     => 1,
+            TN           => $TicketNumber,
+            Title        => $TitleRandom,
+            Queue        => 'Raw',
+            Lock         => 'unlock',
+            Priority     => '3 normal',
+            State        => 'open',
+            CustomerID   => $TestCustomerUserID{UserCustomerID},
+            CustomerUser => $TestCustomerUser,
+            OwnerID      => 1,
+            UserID       => 1,
         );
         $Self->True(
             $TicketID,
             "Ticket is created - ID $TicketID",
+        );
+
+        # add article
+        my $ArticleID = $TicketObject->ArticleCreate(
+            TicketID    => $TicketID,
+            ArticleType => 'email-external',                     # email-external|email-internal|phone|fax|...
+            SenderType  => 'customer',                           # agent|system|customer
+            Subject     => 'First article',                      # required
+            Body        => 'the message text',                   # required
+            ContentType => 'text/plain; charset=ISO-8859-15',    # or optional Charset & MimeType
+            HistoryType    => 'EmailCustomer',     # EmailCustomer|Move|AddNote|PriorityUpdate|WebRequestCustomer|...
+            HistoryComment => 'Some free text!',
+            UserID         => 1,
+        );
+        $Self->True(
+            $ArticleID,
+            "Article is created - ID $ArticleID",
         );
 
         # get script alias
@@ -103,6 +140,20 @@ $Selenium->RunTest(
             $Element->is_enabled();
             $Element->is_displayed();
         }
+
+        # Try to click on the email (link) that should open a popup window.
+        $Selenium->find_element( ".SidebarColumn .WidgetSimple:nth-of-type(2) a.AsPopup", "css" )->VerifiedClick();
+
+        # Wait for popup and switch.
+        $Selenium->WaitFor( WindowCount => 2 );
+        my $Handles = $Selenium->get_window_handles();
+        $Selenium->switch_to_window( $Handles->[1] );
+
+        # wait until page has loaded, if necessary
+        $Selenium->WaitFor( JavaScript => 'return typeof($) === "function" && $("a.UndoClosePopup").length' );
+
+        # close note pop-up window
+        $Selenium->close();
 
         # clean up test data from the DB
         my $Success = $TicketObject->TicketDelete(